+++ src/form-object/form-object.ts
@@ -6,18 +6,17 @@ import { ModelMetadata } from 'types/model-metadata.type';
 import { capitalize } from '../helpers/helpers';
 import { FormObjectOptions } from '../interfaces/form-object-options.interface';
 import { FormGroupOptions } from '../interfaces/form-group-options.interface';
-import { FormModel } from '../interfaces/form-model.interface';
 import { FormStore } from '../form-store/form-store';
 import { ExtendedFormControl } from '../extended-form-control/extended-form-control';
 import { FormError } from './../interfaces/form-error.interface';
 
 // TODO better default values
 const defaultModelOptions: FormObjectOptions = {
-  getConfig: null, // (model: FormModel) => model.config, // TODO see if getConfig can be removed
-  getModelType: (model: FormModel) => model.constructor.name
+  getConfig: null,
+  getModelType: <T>(model: T) => model.constructor.name
 };
 
-export class FormObject {
+export class FormObject<T> {
   protected serviceMappings: object;
 
   public _options: FormObjectOptions;
@@ -25,16 +24,16 @@ export class FormObject {
   public formGroupOptions: FormGroupOptions = {};
   public formStoreClass: any;
 
-  protected beforeSave(store: FormStore): Observable<FormStore> {
+  protected beforeSave(store: FormStore<T>): Observable<FormStore<T>> {
     return observableOf(store);
   }
 
-  protected afterSave(model?: FormModel, form?: FormStore): Observable<FormModel> {
+  protected afterSave(model?: T, form?: FormStore<T>): Observable<T> {
     return observableOf(model);
   }
 
   constructor(
-    public model: FormModel,
+    public model: T,
     protected options: FormObjectOptions
   ) {
     this._options = {
@@ -58,7 +57,7 @@ export class FormObject {
     return modelMetadata.belongsToProperties || [];
   }
 
-  getModelType(model: FormModel): string {
+  getModelType(model: T): string {
     if (this._options.getConfig) {
       // TODO see if can be removed
       return this._options.getConfig(this.model.constructor).type;
@@ -113,14 +112,14 @@ export class FormObject {
     });
   }
 
-  isFormValid(form: FormStore): boolean {
+  isFormValid(form: FormStore<T>): boolean {
     return form.valid || form.disabled;
   }
 
-  public save(form: FormStore): Observable<FormModel> {
+  public save(form: FormStore<T>): Observable<T> {
     return observableOf(true).pipe(
       flatMap(() => this._beforeSave(form)),
-      flatMap((validFormStore: FormStore) => {
+      flatMap((validFormStore: FormStore<T>) => {
         const validatedFormWithModel = new ReplaySubject();
 
         this._save(validFormStore)
@@ -130,7 +129,7 @@ export class FormObject {
               return throwError(error);
             })
           )
-          .subscribe((savedModel: FormModel) => {
+          .subscribe((savedModel: T) => {
             validatedFormWithModel.next({
               savedModel,
               validFormStore
@@ -182,9 +181,9 @@ export class FormObject {
     });
   }
 
-  private _beforeSave(form: FormStore): Observable<FormStore> {
-    const form$: Observable<FormStore> = this.beforeSave(form).pipe(
-      flatMap((transformedForm: FormStore) => {
+  private _beforeSave(form: FormStore<T>): Observable<FormStore<T>> {
+    const form$: Observable<FormStore<T>> = this.beforeSave(form).pipe(
+      flatMap((transformedForm: FormStore<T>) => {
         this.mapPropertiesToModel(transformedForm);
         this.mapBelongsToPropertiesToModel(transformedForm);
 
@@ -199,8 +198,8 @@ export class FormObject {
     return form$;
   }
 
-  private _save(form: FormStore): Observable<FormModel> {
-    const model$: Subject<FormModel> = new Subject<FormModel>();
+  private _save(form: FormStore<T>): Observable<T> {
+    const model$: Subject<T> = new Subject<T>();
 
     const modelType: string = this.getModelType(this.model);
     const service = this.serviceMappings[modelType];
@@ -216,7 +215,7 @@ export class FormObject {
     // issue: if .save() returns BehaviourSubject (which return a value immedietely)
     // .next will be called before "return model$"
     setTimeout(() => {
-      service.save(this.model).subscribe((model: FormModel) => {
+      service.save(this.model).subscribe((model: T) => {
         model$.next(model);
       }, (error: any) => {
         model$.error(error);
@@ -226,9 +225,9 @@ export class FormObject {
     return model$;
   }
 
-  private _afterSave(model: FormModel, form: FormStore): Observable<FormModel> {
-    const form$: Observable<FormModel> = this.afterSave(model, form).pipe(
-      flatMap((transformedModel: FormModel) => {
+  private _afterSave(model: T, form: FormStore<T>): Observable<T> {
+    const form$: Observable<T> = this.afterSave(model, form).pipe(
+      flatMap((transformedModel: T) => {
         this.mapModelPropertiesToForm(transformedModel, form);
         this.resetBelongsToFormControls(transformedModel, form);
         return observableOf(transformedModel);
@@ -239,8 +238,8 @@ export class FormObject {
   }
 
   private mapModelPropertiesToForm(
-    model: FormModel,
-    form: FormStore
+    model: T,
+    form: FormStore<T>
   ): void {
     this.attributeProperties.forEach((propertyName: string) => {
       const formControl: ExtendedFormControl = form.controls[propertyName] as ExtendedFormControl;
@@ -254,7 +253,7 @@ export class FormObject {
     });
   }
 
-  private resetBelongsToFormControls(model: FormModel, form: FormStore): void {
+  private resetBelongsToFormControls(model: T, form: FormStore<T>): void {
     this.belongsToProperties.forEach((propertyName: string) => {
       const formControl: ExtendedFormControl = form.controls[propertyName] as ExtendedFormControl;
       if (formControl.resetValue) {