+++ src/interfaces/form-object-options.interface.ts
@@ -1,5 +1,3 @@
-// import { FormModel } from './form-model.interface';
-
 export interface FormObjectOptions {
   getModelType?: Function;
   getConfig?: Function;